system/requirements.txt
aiohttp>=3.9.0
aiosqlite>=0.20.0
apscheduler>=3.10.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
jieba>=0.42.0
python-dotenv>=1.0.0
simhash>=2.0.0
trafilatura>=1.12.0
fastapi>=0.109.0
uvicorn>=0.27.0
readability-lxml>=0.8.0
pyyaml>=6.0
